+++ quickmin_step/quickmin_parameters.py
@@ -92,6 +92,17 @@ class QuickMinParameters(seamm.Parameters):
             "description": "Forcefield:",
             "help_text": "The forcefield to use.",
         },
+        "calculation": {
+            "default": "optimization",
+            "kind": "enum",
+            "enumeration": (
+                "optimization",
+                "single-point energy",
+            ),
+            "format_string": "",
+            "description": "Calculation:",
+            "help_text": "The type of calculation to perform.",
+        },
         "n_steps": {
             "default": 1000,
             "kind": "integer",
@@ -101,37 +112,6 @@ class QuickMinParameters(seamm.Parameters):
             "description": "Maximum steps:",
             "help_text": "The maximum number of steps to run.",
         },
-        # Put in the configuration handling options needed
-        # "structure handling": {
-        #     "default": "Create a new configuration",
-        #     "kind": "enum",
-        #     "default_units": "",
-        #     "enumeration": (
-        #         "Overwrite the current configuration",
-        #         "Create a new configuration",
-        #     ),
-        #     "format_string": "s",
-        #     "description": "Configuration handling:",
-        #     "help_text": (
-        #         "Whether to overwrite the current configuration, or create a new "
-        #         "configuration or system and configuration for the new structure"
-        #     ),
-        # },
-        # "configuration name": {
-        #     "default": "optimized with <Forcefield>",
-        #     "kind": "string",
-        #     "default_units": "",
-        #     "enumeration": (
-        #         "optimized with <Forcefield>",
-        #         "keep current name",
-        #         "use SMILES string",
-        #         "use Canonical SMILES string",
-        #         "use configuration number",
-        #     ),
-        #     "format_string": "s",
-        #     "description": "Configuration name:",
-        #     "help_text": "The name for the new configuration",
-        # },
         # Results handling
         "results": {
             "default": {},